+++ src/main/java/com/iog/Commands/Music/ClearQueue.java
@@ -7,6 +7,8 @@ import discord4j.common.util.Snowflake;
 import discord4j.core.GatewayDiscordClient;
 import discord4j.core.event.domain.interaction.ChatInputInteractionEvent;
 import discord4j.core.object.VoiceState;
+import discord4j.core.object.command.ApplicationCommand;
+import discord4j.core.object.command.ApplicationCommandOption;
 import discord4j.core.object.entity.Member;
 import discord4j.core.object.entity.Message;
 import discord4j.discordjson.json.ApplicationCommandRequest;
@@ -18,6 +20,7 @@ public class ClearQueue extends BaseCommand {
 		super(
 			new String[]{"clearqueue", "clear"},
 			ApplicationCommandRequest.builder()
+				.type(ApplicationCommand.Type.CHAT_INPUT.getValue())
 				.name("clear-queue")
 				.description("Clears the track queue")
 				.build()