+++ src/main/java/com/iog/Commands/Music/Info.java
@@ -9,6 +9,7 @@ import com.sedmelluq.discord.lavaplayer.source.spotify.SpotifyAudioTrack;
 import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
 import discord4j.common.util.Snowflake;
 import discord4j.core.event.domain.interaction.ChatInputInteractionEvent;
+import discord4j.core.object.command.ApplicationCommand;
 import discord4j.core.object.entity.Member;
 import discord4j.core.object.entity.Message;
 import discord4j.core.spec.EmbedCreateSpec;
@@ -20,6 +21,7 @@ public class Info extends BaseCommand {
 		super(
 			new String[]{"info"},
 			ApplicationCommandRequest.builder()
+				.type(ApplicationCommand.Type.CHAT_INPUT.getValue())
 				.name("info")
 				.description("Gives information about the track playing")
 				.build()