+++ src/main/java/com/iog/Commands/Music/Lyrics.java
@@ -8,6 +8,7 @@ import com.sedmelluq.discord.lavaplayer.track.lyrics.LyricsInfo;
 import com.sedmelluq.discord.lavaplayer.track.lyrics.LyricsManager;
 import discord4j.common.util.Snowflake;
 import discord4j.core.event.domain.interaction.ChatInputInteractionEvent;
+import discord4j.core.object.command.ApplicationCommand;
 import discord4j.core.object.command.ApplicationCommandOption;
 import discord4j.core.object.entity.Message;
 import discord4j.discordjson.json.ApplicationCommandOptionData;
@@ -23,6 +24,7 @@ public class Lyrics extends BaseCommand {
 		super(
 			new String[]{"lyrics"},
 			ApplicationCommandRequest.builder()
+				.type(ApplicationCommand.Type.CHAT_INPUT.getValue())
 				.name("lyrics")
 				.description("Finds lyrics for a song")
 				.addOption(ApplicationCommandOptionData.builder()