+++ src/main/java/com/iog/Commands/Music/Pause.java
@@ -8,6 +8,7 @@ import discord4j.common.util.Snowflake;
 import discord4j.core.GatewayDiscordClient;
 import discord4j.core.event.domain.interaction.ChatInputInteractionEvent;
 import discord4j.core.object.VoiceState;
+import discord4j.core.object.command.ApplicationCommand;
 import discord4j.core.object.command.ApplicationCommandOption;
 import discord4j.core.object.entity.Member;
 import discord4j.core.object.entity.Message;
@@ -21,6 +22,7 @@ public class Pause extends BaseCommand {
 		super(
 			new String[]{"pause"},
 			ApplicationCommandRequest.builder()
+				.type(ApplicationCommand.Type.CHAT_INPUT.getValue())
 				.name("pause")
 				.description("Pauses the current track")
 				.build()