+++ src/main/java/com/iog/Commands/Music/Queue.java
@@ -7,6 +7,7 @@ import com.iog.Utils.CommandExecutionException;
 import com.iog.Utils.Format;
 import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
 import discord4j.core.event.domain.interaction.ChatInputInteractionEvent;
+import discord4j.core.object.command.ApplicationCommand;
 import discord4j.core.object.entity.Message;
 import discord4j.discordjson.json.ApplicationCommandRequest;
 
@@ -19,6 +20,7 @@ public class Queue extends BaseCommand {
 		super(
 			new String[]{"queue", "q"},
 			ApplicationCommandRequest.builder()
+				.type(ApplicationCommand.Type.CHAT_INPUT.getValue())
 				.name("queue")
 				.description("Show the queue in this guild")
 				.build()