+++ src/main/java/com/iog/Commands/Music/Remove.java
@@ -11,6 +11,7 @@ import discord4j.common.util.Snowflake;
 import discord4j.core.GatewayDiscordClient;
 import discord4j.core.event.domain.interaction.ChatInputInteractionEvent;
 import discord4j.core.object.VoiceState;
+import discord4j.core.object.command.ApplicationCommand;
 import discord4j.core.object.command.ApplicationCommandOption;
 import discord4j.core.object.entity.Member;
 import discord4j.core.object.entity.Message;
@@ -24,6 +25,7 @@ public class Remove extends BaseCommand {
 		super(
 			new String[]{"remove", "rm"},
 			ApplicationCommandRequest.builder()
+				.type(ApplicationCommand.Type.CHAT_INPUT.getValue())
 				.name("remove")
 				.description("Removes the track at that the given index")
 				.addOption(ApplicationCommandOptionData.builder()