+++ src/main/java/com/iog/Commands/Music/Resume.java
@@ -8,6 +8,7 @@ import discord4j.common.util.Snowflake;
 import discord4j.core.GatewayDiscordClient;
 import discord4j.core.event.domain.interaction.ChatInputInteractionEvent;
 import discord4j.core.object.VoiceState;
+import discord4j.core.object.command.ApplicationCommand;
 import discord4j.core.object.entity.Member;
 import discord4j.core.object.entity.Message;
 import discord4j.discordjson.json.ApplicationCommandRequest;
@@ -19,6 +20,7 @@ public class Resume extends BaseCommand {
 		super(
 			new String[]{"resume"},
 			ApplicationCommandRequest.builder()
+				.type(ApplicationCommand.Type.CHAT_INPUT.getValue())
 				.name("resume")
 				.description("Resumes the current track")
 				.build()