+++ src/main/java/com/iog/Commands/Music/Stop.java
@@ -6,6 +6,7 @@ import com.iog.Utils.CommandExecutionException;
 import com.iog.Utils.ConnectionUtils;
 import discord4j.common.util.Snowflake;
 import discord4j.core.event.domain.interaction.ChatInputInteractionEvent;
+import discord4j.core.object.command.ApplicationCommand;
 import discord4j.core.object.entity.Member;
 import discord4j.core.object.entity.Message;
 import discord4j.discordjson.json.ApplicationCommandRequest;
@@ -16,6 +17,7 @@ public class Stop extends BaseCommand {
 		super(
 			new String[]{"stop"},
 			ApplicationCommandRequest.builder()
+				.type(ApplicationCommand.Type.CHAT_INPUT.getValue())
 				.name("stop")
 				.description("Stop playback")
 				.build()