+++ src/main/java/com/iog/Commands/Music/Volume.java
@@ -9,6 +9,7 @@ import com.iog.Utils.Format;
 import discord4j.common.util.Snowflake;
 import discord4j.core.event.domain.interaction.ChatInputInteractionEvent;
 import discord4j.core.object.VoiceState;
+import discord4j.core.object.command.ApplicationCommand;
 import discord4j.core.object.command.ApplicationCommandOption;
 import discord4j.core.object.entity.Member;
 import discord4j.core.object.entity.Message;
@@ -22,6 +23,7 @@ public class Volume extends BaseCommand {
 		super(
 			new String[]{"volume", "vol", "v"},
 			ApplicationCommandRequest.builder()
+				.type(ApplicationCommand.Type.CHAT_INPUT.getValue())
 				.name("volume")
 				.description("Show the current volume")
 				.addOption(ApplicationCommandOptionData.builder()