+++ src/main/java/com/iog/Commands/Other/Echo.java
@@ -2,6 +2,7 @@ package com.iog.Commands.Other;
 
 import com.iog.Commands.BaseCommand;
 import discord4j.core.event.domain.interaction.ChatInputInteractionEvent;
+import discord4j.core.object.command.ApplicationCommand;
 import discord4j.core.object.command.ApplicationCommandOption;
 import discord4j.core.object.entity.Message;
 import discord4j.discordjson.json.ApplicationCommandOptionData;
@@ -13,6 +14,7 @@ public class Echo extends BaseCommand {
 		super(
 			new String[]{"echo", "e"},
 			ApplicationCommandRequest.builder()
+				.type(ApplicationCommand.Type.CHAT_INPUT.getValue())
 				.name("echo")
 				.description("Echoes your message")
 				.addOption(ApplicationCommandOptionData.builder()